Changed Generator class to Generator module. Stubs working in enigma test.
